+++ src/main/kotlin/com/handheldrainbow/towerControls/TowerInterface.kt
@@ -1,7 +1,7 @@
 package com.handheldrainbow.towerControls
 
 interface TowerInterface {
-    fun spinA(milliseconds: Long)
-    fun spinB(milliseconds: Long)
+    fun spinA(milliseconds: Float)
+    fun spinB(milliseconds: Float)
     fun stop()
 }
